입저신청 로직 변경
